refactor code structure for schema mapper
